+++ frontend/src/constants/stationSearch.ts
@@ -1,0 +1,3 @@
+export const SEARCH_SCOPE =
+  '&scope=stationName&scope=address&scope=speed&scope=latitude&scope=longitude';
+export const MAX_SEARCH_RESULTS = 10;
