+++ frontend/src/mocks/data.ts
@@ -1,8 +1,9 @@
 import { getTypedObjectFromEntries } from '@utils/getTypedObjectFromEntries';
 import { getTypedObjectKeys } from '@utils/getTypedObjectKeys';
 
-import { ENGLISH_DAYS, MAX_SEARCH_RESULTS } from '@constants';
+import { ENGLISH_DAYS } from '@constants';
 import { CHARGER_TYPES, COMPANY_NAME } from '@constants/chargers';
+import { MAX_SEARCH_RESULTS } from '@constants/stationSearch';
 
 import type {
   CapacityType,
